+++ src/Attributes/NestedAttributes/NestedAttribute.php
@@ -2,16 +2,20 @@
 
 namespace Sintattica\Atk\Attributes\NestedAttributes;
 
+use Exception;
 use Sintattica\Atk\Attributes\Attribute;
 use Sintattica\Atk\Core\Tools;
 use Sintattica\Atk\Db\Query;
 
-class NestedAttribute extends Attribute
+class NestedAttribute extends Attribute implements NestedAttributeInterface
 {
 
-    public function __construct($name, $flags = 0)
+    /**
+     * @throws Exception
+     */
+    public function __construct($name, $flags, string $nestedAttributeField)
     {
-        $this->setIsNestedAttribute(true);
+        $this->setNestedAttributeField($nestedAttributeField);
         parent::__construct($name, $flags);
     }
 
@@ -19,7 +23,7 @@ class NestedAttribute extends Attribute
     {
         parent::setForceUpdate($value);
 
-        $this->m_ownerInstance->getAttribute($this->m_ownerInstance->getNestedAttributeField())->setForceUpdate($value);
+        $this->m_ownerInstance->getAttribute($this->getNestedAttributeField())->setForceUpdate($value);
         return $this;
     }
 
@@ -45,7 +49,7 @@ class NestedAttribute extends Attribute
      */
     public function getSearchCondition(Query $query, $table, $value, $searchmode, $fieldname = '')
     {
-        if (!$this->getOwnerInstance()->hasNestedAttribute($this->fieldName())) {
+        if (!$this->getOwnerInstance()->hasNestedAttribute($this->fieldName(), $this->getNestedAttributeField())) {
             return parent::getSearchCondition($query, $table, $value, $searchmode, $fieldname);
         }
 
@@ -99,7 +103,7 @@ class NestedAttribute extends Attribute
             $tableName = $attr->getDb()->quoteIdentifier($table);
         }
 
-        $nestedAttributeFieldName = $attr->getDb()->quoteIdentifier($attr->m_ownerInstance->getNestedAttributeField());
+        $nestedAttributeFieldName = $attr->getDb()->quoteIdentifier($attr->getNestedAttributeField());
         $nestedAttrName = $attr->fieldName();
 
         return "$tableName.$nestedAttributeFieldName->'$.$nestedAttrName'";
@@ -116,7 +120,7 @@ class NestedAttribute extends Attribute
      */
     static public function getOrderByStatementStatic(Attribute $attr, $extra = [], $table = '', $direction = 'ASC')
     {
-        if ($attr->getOwnerInstance()->hasNestedAttribute($attr->fieldName())) {
+        if ($attr->getOwnerInstance()->hasNestedAttribute($attr->fieldName(), $attr->getNestedAttributeField())) {
             $json_query = self::buildJSONExtractValue($attr, $table);
 
             if ($attr->dbFieldType() == 'string' && $attr->getDb()->getForceCaseInsensitive()) {