+++ src/Attributes/NestedAttributes/NestedAttributeInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Sintattica\Atk\Attributes\NestedAttributes;
+
+use Sintattica\Atk\Db\Query;
+
+interface NestedAttributeInterface
+{
+    public function getSearchCondition(Query $query, $table, $value, $searchmode, $fieldname = '');
+}