+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 This document's format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [11.4.0] - 2025-04-23
+### Fixed
+- Atk: fix nested attributes order and search
+
+
+## [11.3.5] - 2025-04-16
+### Fixed
+- Atk: page language now from user
+
+
+## [11.3.4] - 2025-04-16
+### Fixed
+- Atk: added atk_menu param for username navbar link
+
+
+## [11.3.3] - 2025-04-02
+### Fixed
+- Atk: fix warning in extended search
+
+
 ## [11.3.2] - 2025-01-17
 ### Changed
 - Atk: introduced app_name config for access log